add title confidence scoring and improve media search accuracy
Introduced title confidence scoring using a Levenshtein-based algorithm to enhance media search accuracy. Updated `ImdbProvider` and `TmdbProvider` to calculate and include a `matchConfidence` value when mapping media search results. Adjusted `MediaSearchResult` to represent `matchConfidence` as a `Double` for finer granularity. Refined `MediaProcessor` scoring logic to factor in average provider confidence, boosting result consolidation reliability. These changes significantly improve media match precision and overall result quality.

+++ src/commonMain/kotlin/io/github/hospes/plexify/domain/model/MediaSearchResult.kt
@@ -7,7 +7,7 @@ sealed interface MediaSearchResult {
     val tmdbId: String?
     val tvdbId: String?
     val provider: String
-    val matchConfidence: Int
+    val matchConfidence: Double
 
     data class Movie(
         override val title: String,
@@ -16,7 +16,7 @@ sealed interface MediaSearchResult {
         override val tmdbId: String? = null,
         override val tvdbId: String? = null,
         override val provider: String,
-        override val matchConfidence: Int = 0
+        override val matchConfidence: Double = 0.0,
     ) : MediaSearchResult
 
     data class TvShow(
@@ -26,6 +26,6 @@ sealed interface MediaSearchResult {
         override val tmdbId: String? = null,
         override val tvdbId: String? = null,
         override val provider: String,
-        override val matchConfidence: Int = 0
+        override val matchConfidence: Double = 0.0,
     ) : MediaSearchResult
 }
